Global AirDrops now Disabled
After discussions with testers, we encountered the problem of possible division of loot if two groups want to take it. The PVE concept does not allow to come to any agreement and correctly divide which group and how many items should get. The only solution that has been invented at the moment is to use Light Projectiles that call a Personal Plane with loot. If another idea is invented, the system will be returned.